+++ bank-feeds/RELEASES.md
@@ -749,3 +749,13 @@ Based on:
 - [python v3.2.2] bank-feeds
 ### Releases
 - [PyPI v3.2.2] https://pypi.org/project/codat-bankfeeds/3.2.2 - bank-feeds
+
+## 2023-09-29 00:14:38
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0] bank-feeds
+### Releases
+- [PyPI v3.3.0] https://pypi.org/project/codat-bankfeeds/3.3.0 - bank-feeds